// src/redux/posts/enum.ts
enum PostsActions {
  GET_FEED = "GET_FEED",
  GET_POSTS = "GET_POSTS",
  ADD_POST = "ADD_POST",
  DELETE_POST = "DELETE_POST",
  UPDATE_POST = "UPDATE_POST",
  GET_POST_COMMENTS = "GET_POST_COMMENTS",
  ADD_POST_COMMENT = "ADD_POST_COMMENT",
  UPDATE_POST_COMMENT = "UPDATE_POST_COMMENT",
  DELETE_POST_COMMENT = "DELETE_POST_COMMENT",
  ADD_POST_LIKE = "ADD_POST_LIKE",
  DELETE_POST_LIKE = "DELETE_POST_LIKE",
  GET_VIEWED_USER_POSTS = "GET_VIEWED_USER_POSTS",
}

export default PostsActions;